+++ lib/src/numdart/arrays_base/array_operations/array_add_to_scalar.dart
@@ -1,21 +1,21 @@
 import '../array.dart';
 
-/// Subtract all the array elements for a number
+/// Add a number for each array element
 ///
 /// # Examples
 /// ```dart
 /// var list = Array([4.0, 4.0 , 4.0]);
 ///
-/// print(arraySubToScalar(list, 2));
+/// print(arrayAddToScalar(list, 2));
 ///
 /// /* output:
-/// Array([2.0, 2.0, 2.0])
+/// Array([6.0, 6.0, 6.0])
 /// */
 /// ```
-Array arraySubToScalar(Array a, num b) {
+Array arrayAddToScalar(Array a, num b) {
   var c = Array.fixed(a.length);
   for (var i = 0; i < a.length; i++) {
-    c[i] = a[i] - b;
+    c[i] = a[i] + b;
   }
   return c;
 }