+++ test/numdart/arrays_base/array2d_operations/array2d_sub_to_scalar_test.dart
@@ -1,0 +1,24 @@
+import 'package:scidart/src/numdart/numdart.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('subtractEachElement', () {
+    var a = Array2d([
+      Array([2, 2, 2]),
+      Array([2, 2, 2]),
+      Array([2, 2, 2])
+    ]);
+
+    array2dSubToScalar(a, 2);
+
+    print(a);
+
+    var aExpec = Array2d([
+      Array([0, 0, 0]),
+      Array([0, 0, 0]),
+      Array([0, 0, 0])
+    ]);
+
+    expect(a, aExpec);
+  });
+}