+++ test/numdart/arrays_base/array_operations/array_pad_start_test.dart
@@ -1,0 +1,13 @@
+import 'package:scidart/src/numdart/numdart.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('pad array with zeros at start', () {
+    var a = Array([4.0, 4.0, 4.0]);
+    var b = arrayPadStart(a, 2);
+
+    var bExp = Array([0, 0, 4.0, 4.0, 4.0]);
+
+    expect(b, bExp);
+  });
+}